client: better encapsulation of signature data types
